Start opendaylight as systemd service
Karaf container runs as java application and we cannot detect if
karaf is already running if there is any other active java process.
opendaylight rpm package will include .service file for controller.

Changed the logfile place holder for barclamp and replaced it with
a list of features to be installed in opendaylight controller.

I wanted to keep the logfile parameter but as opendaylight runs as
a service, to keep logfile parameter I must add KARAF_REDIRECT environment
variable in opendaylight.service file or add a new .conf file
for opendaylight and include that in opendaylight.service file. Too
much hassle for a single parameter. Will do that in future if we
decide to add more configurable parameter for karaf.