Added common dependency configurations and contrail roles.
Updated the config files for dependency components for contrail control
plane. These are common packages and configurations and may need to be
installed and linked to other contrail supervisor services. The contrail
supervisor services are themselves should be able to be applied
independently of each other to form a cluster. Hence new roles and
correspondig recipe framework is added to be able to assign roles to
multi-node cluster. The recipes are yet to be coded and updated.